+++ scripts/lateralus.mixins.js
@@ -150,6 +150,20 @@ define([
   };
 
   /**
+   * Listen to an event-emitting Object and amplify one of its events across
+   * the {{#crossLink "Lateralus"}}{{/crossLink}} application.  Useful for
+   * making plain Backbone Objects (i.e., non-Lateralus Objects) communicate
+   * important information in a broader way.
+   * @method amplify
+   * @param {Backbone.Events} emitter The object that `trigger`s events that
+   * should be amplified globally across the app.
+   * @param {string} eventName The event to amplify globally across the app.
+   */
+  mixins.amplify = function (emitter, eventName) {
+    this.listenTo(emitter, eventName, _.bind(this.emit, this, eventName));
+  };
+
+  /**
    * Listen for an event that is triggered on the central {{#crossLink
    * "Lateralus"}}{{/crossLink}} instance and bind a function handler.
    * @method listenFor